flatten handler files
