--- mocks/handlers/current-user/index.ts
@@ -1,1 +1,0 @@
-export * from './handler';
