--- mocks/handlers/items/id/children/index.ts
@@ -1,1 +1,0 @@
-export { itemIdChildrenHandler } from './handler';
