--- mocks/handlers/items/id/index.ts
@@ -1,2 +1,0 @@
-export { itemIdHandler } from './handler';
-export * from './children';
